Route scoring seems to be working

+++ src/utils/uuid.cpp
@@ -15,7 +15,7 @@ auto bytes(std::size_t quantity) -> std::string
   std::string output;
   output.reserve(quantity);
 
-  static std::uniform_int_distribution<uint8_t> distribution(std::numeric_limits<uint8_t>::min(), std::numeric_limits<uint8_t>::max());
+  static std::uniform_int_distribution<int> distribution(std::numeric_limits<uint8_t>::min(), std::numeric_limits<uint8_t>::max()); // MSVC doesn't like std::uniform_int_distribution<uint8_t>
 
   // Generate quantity-worth of random bytes and add to the output string.
   for (std::size_t n = 0; n < quantity; ++n)
